+++ src/inmanta/server/services/compilerservice.py
@@ -417,14 +417,25 @@ class CompilerService(ServerSlice):
 
     async def _queue(self, compile: data.Compile) -> None:
         async with self._global_lock:
+            env: Optional[data.Environment] = await data.Environment.get_by_id(compile.environment)
+            if env is None:
+                raise Exception("Can't queue compile: environment %s does not exist" % compile.environment)
+            assert env is not None
+            # don't execute any compiles in a halted environment
+            if env.halted:
+                return
+
             if compile.environment not in self._recompiles or self._recompiles[compile.environment].done():
                 task = self.add_background_task(self._run(compile))
                 self._recompiles[compile.environment] = task
 
     async def _dequeue(self, environment: uuid.UUID) -> None:
         async with self._global_lock:
+            env: Optional[data.Environment] = await data.Environment.get_by_id(environment)
+            if env is None:
+                raise Exception("Can't queue compile: environment %s does not exist" % environment)
             nextrun = await data.Compile.get_next_run(environment)
-            if nextrun:
+            if nextrun and not env.halted:
                 task = self.add_background_task(self._run(nextrun))
                 self._recompiles[environment] = task
             else:
@@ -452,6 +463,14 @@ class CompilerService(ServerSlice):
         unhandled = await data.Compile.get_unhandled_compiles()
         for u in unhandled:
             self.add_background_task(self._notify_listeners(u))
+
+    async def resume_environment(self, environment: uuid.UUID) -> None:
+        """
+            Resume compiler service after halt.
+        """
+        compile: Optional[data.Compile] = await data.Compile.get_next_run(environment)
+        if compile is not None:
+            await self._queue(compile)
 
     @protocol.handle(methods.is_compiling, environment_id="id")
     async def is_compiling(self, environment_id: uuid.UUID) -> Apireturn: